dt-bindings: irq: sun7i-nmi: correct Allwinner A523 compatible
After some testing it became evident that the NMI controller as used in
the Allwinner A523 SoC is NOT compatible to the previous models, as it
requires an extra bit to deliver the IRQs.

Move the A523 compatible string out of the fallback list, and let it
stand on its own.
So far there are no DTs using this compatible committed, so this
incompatible change change should not introduce regressions.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>