remoteproc: qcom_q6v5_pas: Attach running remoteproc if firmware is missing
A remoteproc might be already running during boot, e.g. because it was
already started by the boot firmware. This is the case for example on X1E,
where the boot firmware starts a "lite" ADSP firmware that supports
charging and USB-CC detection, but is missing audio functionality. This
firmware uses the same interfaces as the full firmware and can be reused in
case the device-specific firmware is missing (e.g. in generic distro
installers).

The running remoteproc is currently not modelled at all - it is just killed
through qcom_scm_pas_shutdown() without even using the SMP2P stop signal
beforehand.

Model this properly by marking the remoteproc as RPROC_DETACHED based on
the SMP2P state and use the new RPROC_AUTO_BOOT_RESTART_IF_FW_AVAILABLE
feature to implement equivalent behavior as before. If the firmware is
present the "lite" firmware is now stopped more gracefully with the SMP2P
stop signal. If the firmware is missing, we attach to it and can provide
battery status and USB-C detection until the full ADSP firmware is copied.

Note that the exact same approach does not just work for the "lite" ADSP
firmware. Any of the remoteprocs (ADSP, CDSP, SLPI, ...) can be already
started during boot and attached (or restarted) later.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>